MS C++ warning fixes from Harrie Hazewinkel
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V4-2-patches@5806 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/asn1.c
@@ -162,7 +162,7 @@ asn_check_packet (u_char *pkt, size_t len)
 }
 
 static
-int _asn_bitstring_check(const char * str, u_long asn_length, u_char datum)
+int _asn_bitstring_check(const char * str, size_t asn_length, u_char datum)
 {
     char ebuf[128];
